Added `announce` command and imporvements
Added `announce` command that is able to announce any messaage to any channel in which bot has sendMessages permission. It sends announcement in form of embed, user can change content and title, also there is a option to hide who is the announcer.

Minor change in status.js - added option to choose wheter to display result publicly or not